Move Ball, Paddle, and PaddleController classe/struct definitions out of PongLayer .h/.cpp
