+++ Sandbox/PongLayer.cpp
@@ -1,6 +1,6 @@
 #include "PongLayer.h"
 
-#include "Aquarius/Renderer/Renderer.h"
+#include "Aquarius.h"
 
 #include <array>
 #include <cmath>
@@ -13,130 +13,6 @@ static constexpr auto x = 0;
 static constexpr auto y = 1;
 static constexpr auto pi = 3.14159265358979323846;
 
-void moveUp(Aquarius::timeDelta_t dt, Paddle* paddle)
-{
-	if (paddle->position.y > 0)
-	{
-		auto dy = dt * paddle->speedY;
-		paddle->position.y -= dy;
-	}
-}
-
-void moveDown(Aquarius::timeDelta_t dt, Paddle* paddle)
-{
-	if (paddle->position.y < Aquarius::Application::get()->getWindow()->getHeight() - 80.0)
-	{
-		auto dy = dt * paddle->speedY;
-		paddle->position.y += dy;
-	}
-}
-
-KeyboardPaddleController::KeyboardPaddleController(Aquarius::Input::KeyCode upKey, Aquarius::Input::KeyCode downKey)
-	: m_UpKey(upKey), m_DownKey(downKey)
-{}
-
-void KeyboardPaddleController::movePaddle(Aquarius::timeDelta_t dt, Paddle* paddle, Ball* ball)
-{
-	if (Aquarius::Input::isKeyPressed(m_UpKey))
-	{
-		moveUp(dt, paddle);
-	}
-
-	if (Aquarius::Input::isKeyPressed(m_DownKey))
-	{
-		moveDown(dt, paddle);
-	}
-}
-
-void MousePaddleController::movePaddle(Aquarius::timeDelta_t dt, Paddle* paddle, Ball* ball)
-{
-	if (Aquarius::Input::getMousePosition().y > (paddle->position.y + (paddle->size.y / 2.0)))
-	{
-		moveDown(dt, paddle);
-	}
-
-	if (Aquarius::Input::getMousePosition().y < (paddle->position.y + (paddle->size.y / 2.0)))
-	{
-		moveUp(dt, paddle);
-	}
-}
-
-void AIPaddleController::movePaddle(Aquarius::timeDelta_t dt, Paddle* paddle, Ball* ball)
-{
-	if (paddle->position.y + (paddle->size.y / 2.0) > ball->position.y + (ball->size.y / 2.0))
-	{
-		moveUp(dt, paddle);
-	}
-	else if (paddle->position.y + (paddle->size.y / 2.0) < ball->position.y + (ball->size.y / 2.0))
-	{
-		moveDown(dt, paddle);
-	}
-}
-
-void Paddle::Render() const 
-{
-	Aquarius::Renderer::DrawQuad(
-		position,
-		size,
-		0,
-		{ 1.0, 0.5, 1.0, 1.0 }
-	);
-}
-
-void Ball::Render() const
-{
-	Aquarius::Renderer::DrawQuad(
-		position,
-		size,
-		0,
-		{ 1.0, 0.5, 1.0, 1.0 }
-	);
-}
-
-void Ball::Update(Aquarius::timeDelta_t dt)
-{
-	Aquarius::Application* app = Aquarius::Application::get();
-	Aquarius::Window* window = app->getWindow();
-
-	int height = window->getHeight();
-	int width = window->getWidth();
-	auto dx = dt * velocity.x;
-	auto dy = dt * velocity.y;
-	
-	position.x += dx;
-	position.y += dy;
-
-	if (position.y <= 0 || position.y >= window->getHeight() - 10.0)
-	{
-		position.y -= dy;
-		velocity.y = -velocity.y;
-	}
-
-	if (position.x <= 0)
-	{
-		AQ_INFO("Right Paddle Scored!");
-		Reset(false);
-	}
-
-	if (position.x >= window->getWidth() + 10.0)
-	{
-		AQ_INFO("Left Paddle Scored!");
-		Reset(true);
-	}
-}
-
-void Ball::Reset(bool leftScored)
-{
-	static float yVelocity = 0.2;
-	Aquarius::Application* app = Aquarius::Application::get();
-	Aquarius::Window* window = app->getWindow();
-
-	yVelocity = -yVelocity;
-	// TODO Randomize starting velocity
-	position = { (window->getWidth() / 2.0) - 5.0, (window->getHeight() / 2.0) - 5.0 };
-	velocity = { leftScored ? 0.2 : -0.2, yVelocity };
-	speed = { 0.4, 0.4 };
-}
 
 PongLayer::PongLayer()
 	: Layer("Pong"), m_LeftPaddle(), m_RightPaddle()
